Auth_Code_Flow_Authentication/src/requirements.txt
fastapi 
uvicorn 
msal 
aiohttp
python-dotenv
rich